test: Support printf-tests for 16-bit int targets
Select between correct 32 and 16 bit output values for each test case.
Cast size_t strlen result to int.

+++ test/printf-tests.c
@@ -120,7 +120,7 @@ static int test(int serial, char *expect, char *fmt, ...) {
     }
     if (n != (int) strlen(expect)) {
         failmsg(serial, "expected \"%s\" (%d), got \"%s\" (%d)",
-		expect, strlen(expect), buf, n);
+		expect, (int) strlen(expect), buf, n);
 	free(abuf);
         return 1;
     }
